+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/service/AppService.java
@@ -162,13 +162,6 @@ public class AppService {
       return app;
     }
 
-    UserInfo owner = userService.findByUserId(app.getOwnerName());
-    if (owner == null) {
-      throw new BadRequestException("Application's owner not exist.");
-    }
-
-    app.setOwnerEmail(owner.getEmail());
-
     App createdApp = appRepository.save(app);
 
     roleInitializationService.initAppRoles(createdApp);