+++ CHANGES.md
@@ -42,5 +42,6 @@ Apollo 2.0.0
 * [Fix bug: associated namespace display incorrect in text view](https://github.com/apolloconfig/apollo/pull/4219)
 * [Add Ordered interface to ProviderManager SPI](https://github.com/apolloconfig/apollo/pull/4218)
 * [Using commons-lang3 to replace commons-lang](https://github.com/apolloconfig/apollo/pull/4225)
+* [optimize import/export config](https://github.com/apolloconfig/apollo/pull/4231)
 ------------------
 All issues and pull requests are [here](https://github.com/ctripcorp/apollo/milestone/8?closed=1)